{"schema":"decodex/commit/1","summary":"Split quantitative ranking query detection","authority":"manual"}
