+++ apps/elf-eval/src/bin/real_world_job_benchmark/quantitative/metrics/ranking.rs
@@ -1,9 +1,11 @@
-use crate::{BTreeMap, BTreeSet, RealWorldJob, ReportSummary, scoring};
+mod queries;
+
+use crate::{BTreeSet, RealWorldJob, ReportSummary};
 
 pub(super) fn ranking_query_ids(source_jobs: &[RealWorldJob]) -> BTreeSet<&str> {
 	source_jobs
 		.iter()
-		.filter(|job| !ranking_relevance_grades(job).is_empty() && ranking_query_attempted(job))
+		.filter(|job| queries::is_ranking_query(job))
 		.map(|job| job.job_id.as_str())
 		.collect()
 }
@@ -47,37 +49,4 @@ pub(super) fn ranking_coverage_state(
 
 pub(super) fn ranked_candidate_source(ranking_query_count: usize) -> &'static str {
 	if ranking_query_count == 0 { "not_encoded" } else { "produced_evidence_order" }
-}
-
-fn ranking_relevance_grades(source_job: &RealWorldJob) -> BTreeMap<String, f64> {
-	if !source_job.expected_answer.relevance_judgments.is_empty() {
-		return source_job
-			.expected_answer
-			.relevance_judgments
-			.iter()
-			.filter(|judgment| judgment.grade > 0.0)
-			.map(|judgment| (judgment.evidence_id.clone(), judgment.grade))
-			.collect();
-	}
-
-	source_job
-		.required_evidence
-		.iter()
-		.filter(|evidence| matches!(evidence.requirement.as_str(), "cite" | "use" | "explain"))
-		.map(|evidence| (evidence.evidence_id.clone(), 1.0))
-		.collect()
-}
-
-fn ranking_query_attempted(job: &RealWorldJob) -> bool {
-	if !scoring::produced_evidence_order(job).is_empty() {
-		return true;
-	}
-
-	let Some(answer) = job.corpus.adapter_response.as_ref().map(|response| &response.answer) else {
-		return false;
-	};
-
-	answer.trace_explainability.as_ref().is_some_and(|trace| {
-		trace.stages.iter().any(|stage| stage.stage_name == "live_adapter.retrieve")
-	}) && answer.latency_ms.is_some_and(|latency| latency.is_finite() && latency > 0.0)
 }
